backend/requirements.txt
fastapi
uvicorn[standard]
websockets
autogen-core
autogen-agentchat
autogen-ext
psycopg2-binary
asyncpg
python-dotenv
paramiko
pandas
numpy
torch
transformers
scikit-learn
python-docx
python-dateutil
requests
colorama
typing-extensions
websockets>=11.0.3
asyncio-mqtt>=0.11.1 
openai
tiktoken